Move default narrator hydration into narrator store
